third_party/WebKit/LayoutTests/http/tests/websocket/multiple-connections-throttled-expected.txt

CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
CONSOLE ERROR: line 24: WebSocket connection to 'ws://127.0.0.1:8880/echo' failed: Error in connection establishment: net::ERR_INSUFFICIENT_RESOURCES
Test that WebSocket connections are throttled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS openedCount is 255
PASS successfullyParsed is true

TEST COMPLETE
